# api/spyglass.openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "Infinite Ocean: Spyglass API", "version": "0.1.0", "description": "Customer-facing JSON API. Cell-owned operations are reached through the authenticated Spyglass router; clients never select a cell."},
  "servers": [{"url": "https://app.infiniteocean.net"}],
  "paths": {
...
    "/api/v1/accounts/{accountID}/context": {"parameters": [{"$ref": "#/components/parameters/accountID"}], "get": {"operationId": "accountContext", "x-spyglass-service": "cell-api", "security": [{"sessionCookie": []}], "responses": {"200": {"$ref": "#/components/responses/AccountContextOK"}, "default": {"$ref": "#/components/responses/Problem"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "__Host-spyglass_session"},
      "stripeSignature": {"type": "apiKey", "in": "header", "name": "Stripe-Signature"},
      "exportDownloadCapability": {"type": "apiKey", "in": "header", "name": "Authorization", "description": "Short-lived SPYGLASS-ACCOUNT-EXPORT capability; never place it in a URL"}
    },
    "parameters": {
      "accountID": {"name": "accountID", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
    },
    "schemas": {
      "Problem": {"type": "object", "required": ["type", "title", "status", "code", "detail"], "properties": {"type": {"type": "string", "format": "uri", "pattern": "^https://infiniteocean\\.net/problems/[a-z][a-z0-9_]{0,99}$"}, "title": {"type": "string", "minLength": 1}, "status": {"type": "integer", "minimum": 400, "maximum": 599}, "code": {"type": "string", "pattern": "^[a-z][a-z0-9_]{0,99}$"}, "detail": {"type": "string", "minLength": 1}}, "additionalProperties": false},
...
      "PackageAccess": {"type": "object", "required": ["code", "version", "mode"], "properties": {"code": {"type": "string", "pattern": "^[a-z][a-z0-9_-]{0,49}$"}, "version": {"type": "integer", "minimum": 1}, "mode": {"type": "string", "enum": ["enabled", "read_only", "suspended"]}, "limits": {"type": "object", "additionalProperties": {"type": "integer", "minimum": 0}}, "limit_policies": {"type": "object", "additionalProperties": {"$ref": "#/components/schemas/LimitPolicy"}}}, "additionalProperties": false},
      "LimitPolicy": {"type": "object", "required": ["kind", "combine"], "properties": {"kind": {"type": "string", "enum": ["capacity"]}, "combine": {"type": "string", "enum": ["replace", "add", "maximum", "minimum"]}, "reservation_ttl_seconds": {"type": "integer", "minimum": 1}}, "additionalProperties": false},
      "AccountContext": {"type": "object", "required": ["account_id", "actor_kind", "role", "cell_id", "placement_generation", "entitlement_version", "package_access"], "properties": {"account_id": {"type": "string", "format": "uuid"}, "actor_kind": {"type": "string", "enum": ["user", "workload"]}, "role": {"type": "string", "enum": ["owner", "administrator", "billing_admin", "member", "viewer"]}, "cell_id": {"type": "string", "pattern": "^[a-z][a-z0-9-]{0,63}$"}, "placement_generation": {"type": "integer", "minimum": 1}, "entitlement_version": {"type": "integer", "minimum": 1}, "package_access": {"anyOf": [{"$ref": "#/components/schemas/PackageAccess"}, {"type": "null"}]}}, "additionalProperties": false},
...
    },
...
    "responses": {
...
      "AccountContextOK": {"description": "Trusted routed Account context", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountContext"}}}},
...
      "Problem": {"description": "Request failed", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    }
  }
}